+++ wire-ios-share-engine/Sources/SharingSession.swift
@@ -24,144 +24,6 @@ import WireNetwork
 import WireRequestStrategy
 import WireTransport
 
-final class PushMessageHandlerDummy: NSObject, PushMessageHandler {
-
-    func didFailToSend(_ message: ZMMessage) {
-        // nop
-    }
-}
-
-final class ClientRegistrationStatus: NSObject, ClientRegistrationDelegate {
-
-    let context: NSManagedObjectContext
-
-    init(context: NSManagedObjectContext) {
-        self.context = context
-    }
-
-    var clientIsReadyForRequests: Bool {
-        // swiftlint:disable:next todo_requires_jira_link
-        // TODO: move constant into shared framework
-        if let clientId = context.persistentStoreMetadata(forKey: ZMPersistedClientIdKey) as? String {
-            return !clientId.isEmpty
-        }
-
-        return false
-    }
-
-    func didDetectCurrentClientDeletion() {
-        // nop
-    }
-}
-
-final class AuthenticationStatus: AuthenticationStatusProvider {
-
-    let transportSession: ZMTransportSession
-
-    init(transportSession: ZMTransportSession) {
-        self.transportSession = transportSession
-    }
-
-    var state: AuthenticationState {
-        isLoggedIn ? .authenticated : .unauthenticated
-    }
-
-    private var isLoggedIn: Bool {
-        transportSession.cookieStorage.hasAuthenticationCookie
-    }
-
-}
-
-extension BackendEnvironmentProvider {
-    func cookieStorage(for account: Account) -> ZMPersistentCookieStorage {
-        let backendURL = backendURL.host!
-        return ZMPersistentCookieStorage(
-            forServerName: backendURL,
-            userIdentifier: account.userIdentifier,
-            useCache: false
-        )
-    }
-
-    public func isAuthenticated(_ account: Account) -> Bool {
-        cookieStorage(for: account).hasAuthenticationCookie
-    }
-}
-
-final class ApplicationStatusDirectory: ApplicationStatus {
-
-    let transportSession: ZMTransportSession
-
-    /// The authentication status used to verify a user is authenticated
-    public let authenticationStatus: AuthenticationStatusProvider
-
-    /// The client registration status used to lookup if a user has registered a self client
-    public let clientRegistrationStatus: ClientRegistrationDelegate
-
-    public let linkPreviewDetector: LinkPreviewDetectorType
-
-    public let syncStatus: SyncStatusProtocol
-
-    public init(
-        transportSession: ZMTransportSession,
-        authenticationStatus: AuthenticationStatusProvider,
-        clientRegistrationStatus: ClientRegistrationStatus,
-        linkPreviewDetector: LinkPreviewDetectorType,
-        syncStatus: SyncStatusProtocol = SyncStatus()
-    ) {
-        self.transportSession = transportSession
-        self.authenticationStatus = authenticationStatus
-        self.clientRegistrationStatus = clientRegistrationStatus
-        self.linkPreviewDetector = linkPreviewDetector
-        self.syncStatus = syncStatus
-    }
-
-    public convenience init(syncContext: NSManagedObjectContext, transportSession: ZMTransportSession) {
-        let authenticationStatus = AuthenticationStatus(transportSession: transportSession)
-        let clientRegistrationStatus = ClientRegistrationStatus(context: syncContext)
-        let linkPreviewDetector = LinkPreviewDetector()
-        self.init(
-            transportSession: transportSession,
-            authenticationStatus: authenticationStatus,
-            clientRegistrationStatus: clientRegistrationStatus,
-            linkPreviewDetector: linkPreviewDetector
-        )
-    }
-
-    public var synchronizationState: SynchronizationState {
-        if clientRegistrationStatus.clientIsReadyForRequests {
-            .online
-        } else {
-            .unauthenticated
-        }
-    }
-
-    public var operationState: OperationState {
-        .foreground
-    }
-
-    public var clientRegistrationDelegate: ClientRegistrationDelegate {
-        clientRegistrationStatus
-    }
-
-    public var requestCancellation: ZMRequestCancellation {
-        transportSession
-    }
-
-    func requestResyncResources() {
-        // we don't resync Resources in the share engine
-    }
-
-}
-
-/// Required by `MLSService` initializer.
-/// No need to fill in the methods as we don't sync resources in the share engine.
-struct SyncStatus: SyncStatusProtocol {
-    func performQuickSync() async {}
-    func resyncResources() {}
-    func forceSlowSync() {}
-    func recoverWithQuickSync() async {}
-    var isLive: Bool = false
-}
 
 /// A Wire session to share content from a share extension
 /// - note: this is the entry point of this framework. Users of
